+++ sdk/purview/purview-administration-rest/samples-dev/metadataPolicies.ts
@@ -8,7 +8,10 @@
  * @azsdk-weight 40
  */
 
-import { MetadataPolicies } from "@azure-rest/purview-administration";
+import {
+  PurviewMetadataPolicies,
+  PurviewMetadataPoliciesClient,
+} from "@azure-rest/purview-administration";
 import { DefaultAzureCredential } from "@azure/identity";
 import dotenv from "dotenv";
 
@@ -18,10 +21,7 @@ const endpoint = process.env["ENDPOINT"] || "";
 
 async function main() {
   console.log("== List metadata policies sample ==");
-  const client = MetadataPolicies.PurviewMetadataPoliciesClient(
-    endpoint,
-    new DefaultAzureCredential()
-  );
+  const client = PurviewMetadataPoliciesClient(endpoint, new DefaultAzureCredential());
 
   const response = await client.path("/metadataPolicies").get();
 
@@ -31,7 +31,7 @@ async function main() {
     throw new Error(error);
   }
 
-  const policies = MetadataPolicies.paginate(client, response);
+  const policies = PurviewMetadataPolicies.Pagination.paginate(client, response);
 
   for await (const policy of policies) {
     console.log(policy.name);
