+++ sdk/purview/purview-administration-rest/src/account/index.ts
@@ -1,123 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
-import * as models from "./models";
-import * as paginateHelper from "./paginateHelper";
-import * as parameters from "./parameters";
-import * as purviewAccount from "./purviewAccount";
-import * as responses from "./responses";
 
-export namespace Account {
-  export type Account = models.Account;
-  export type Identity = models.Identity;
-  export type AccountProperties = models.AccountProperties;
-  export type CloudConnectors = models.CloudConnectors;
-  export type AccountEndpoints = models.AccountEndpoints;
-  export type AccountPropertiesEndpoints = models.AccountPropertiesEndpoints;
-  export type ManagedResources = models.ManagedResources;
-  export type AccountPropertiesManagedResources = models.AccountPropertiesManagedResources;
-  export type PrivateEndpointConnection = models.PrivateEndpointConnection;
-  export type PrivateEndpointConnectionProperties = models.PrivateEndpointConnectionProperties;
-  export type PrivateEndpoint = models.PrivateEndpoint;
-  export type PrivateLinkServiceConnectionState = models.PrivateLinkServiceConnectionState;
-  export type AccountSku = models.AccountSku;
-  export type SystemData = models.SystemData;
-  export type AccountSystemData = models.AccountSystemData;
-  export type ErrorResponseModel = models.ErrorResponseModel;
-  export type ErrorModel = models.ErrorModel;
-  export type ErrorResponseModelError = models.ErrorResponseModelError;
-  export type DataPlaneAccountUpdateParameters = models.DataPlaneAccountUpdateParameters;
-  export type AccessKeys = models.AccessKeys;
-  export type AccessKeyOptions = models.AccessKeyOptions;
-  export type Collection = models.Collection;
-  export type CollectionReference = models.CollectionReference;
-  export type CollectionSystemData = models.CollectionSystemData;
-  export type CollectionList = models.CollectionList;
-  export type CollectionNameResponseList = models.CollectionNameResponseList;
-  export type CollectionNameResponse = models.CollectionNameResponse;
-  export type CollectionPathResponse = models.CollectionPathResponse;
-  export type ResourceSetRuleConfig = models.ResourceSetRuleConfig;
-  export type AdvancedResourceSet = models.AdvancedResourceSet;
-  export type PathPatternExtractorConfig = models.PathPatternExtractorConfig;
-  export type Filter = models.Filter;
-  export type ComplexReplacerConfig = models.ComplexReplacerConfig;
-  export type NormalizationRule = models.NormalizationRule;
-  export type FastRegex = models.FastRegex;
-  export type RegexReplacer = models.RegexReplacer;
-  export type ScopedRule = models.ScopedRule;
-  export type Rule = models.Rule;
-  export type ResourceSetRuleConfigList = models.ResourceSetRuleConfigList;
-
-  export type GetArrayType<T> = paginateHelper.GetArrayType<T>;
-  export type GetPage<T> = paginateHelper.GetPage<T>;
-  export type PaginateReturn<T> = paginateHelper.PaginateReturn<T>;
-  export type PagingOptions<T> = paginateHelper.PagingOptions<T>;
-  export const paginate = paginateHelper.paginate;
-
-  export type AccountsGetAccountPropertiesParameters = parameters.AccountsGetAccountPropertiesParameters;
-  export type AccountsUpdateAccountPropertiesParameters = parameters.AccountsUpdateAccountPropertiesParameters;
-  export type AccountsGetAccessKeysParameters = parameters.AccountsGetAccessKeysParameters;
-  export type AccountsRegenerateAccessKeyParameters = parameters.AccountsRegenerateAccessKeyParameters;
-  export type CollectionsGetCollectionParameters = parameters.CollectionsGetCollectionParameters;
-  export type CollectionsCreateOrUpdateCollectionParameters = parameters.CollectionsCreateOrUpdateCollectionParameters;
-  export type CollectionsDeleteCollectionParameters = parameters.CollectionsDeleteCollectionParameters;
-  export type CollectionsListCollectionsParameters = parameters.CollectionsListCollectionsParameters;
-  export type CollectionsListChildCollectionNamesParameters = parameters.CollectionsListChildCollectionNamesParameters;
-  export type CollectionsGetCollectionPathParameters = parameters.CollectionsGetCollectionPathParameters;
-  export type ResourceSetRulesGetResourceSetRuleParameters = parameters.ResourceSetRulesGetResourceSetRuleParameters;
-  export type ResourceSetRulesCreateOrUpdateResourceSetRuleParameters = parameters.ResourceSetRulesCreateOrUpdateResourceSetRuleParameters;
-  export type ResourceSetRulesDeleteResourceSetRuleParameters = parameters.ResourceSetRulesDeleteResourceSetRuleParameters;
-  export type ResourceSetRulesListResourceSetRulesParameters = parameters.ResourceSetRulesListResourceSetRulesParameters;
-  export type AccountsUpdateAccountPropertiesBodyParam = parameters.AccountsUpdateAccountPropertiesBodyParam;
-  export type AccountsRegenerateAccessKeyBodyParam = parameters.AccountsRegenerateAccessKeyBodyParam;
-  export type CollectionsCreateOrUpdateCollectionBodyParam = parameters.CollectionsCreateOrUpdateCollectionBodyParam;
-  export type CollectionsListCollectionsQueryParamProperties = parameters.CollectionsListCollectionsQueryParamProperties;
-  export type CollectionsListCollectionsQueryParam = parameters.CollectionsListCollectionsQueryParam;
-  export type CollectionsListChildCollectionNamesQueryParamProperties = parameters.CollectionsListChildCollectionNamesQueryParamProperties;
-  export type CollectionsListChildCollectionNamesQueryParam = parameters.CollectionsListChildCollectionNamesQueryParam;
-  export type ResourceSetRulesCreateOrUpdateResourceSetRuleBodyParam = parameters.ResourceSetRulesCreateOrUpdateResourceSetRuleBodyParam;
-  export type ResourceSetRulesListResourceSetRulesQueryParamProperties = parameters.ResourceSetRulesListResourceSetRulesQueryParamProperties;
-  export type ResourceSetRulesListResourceSetRulesQueryParam = parameters.ResourceSetRulesListResourceSetRulesQueryParam;
-
-  export const PurviewAccountClient = purviewAccount.PurviewAccount;
-  export type PurviewAccountClientLike = purviewAccount.PurviewAccountRestClient;
-  export type AccountsGetAccountProperties = purviewAccount.AccountsGetAccountProperties;
-  export type AccountsGetAccessKeys = purviewAccount.AccountsGetAccessKeys;
-  export type AccountsRegenerateAccessKey = purviewAccount.AccountsRegenerateAccessKey;
-  export type CollectionsGetCollection = purviewAccount.CollectionsGetCollection;
-  export type CollectionsListCollections = purviewAccount.CollectionsListCollections;
-  export type CollectionsListChildCollectionNames = purviewAccount.CollectionsListChildCollectionNames;
-  export type CollectionsGetCollectionPath = purviewAccount.CollectionsGetCollectionPath;
-  export type ResourceSetRulesGetResourceSetRule = purviewAccount.ResourceSetRulesGetResourceSetRule;
-  export type ResourceSetRulesListResourceSetRules = purviewAccount.ResourceSetRulesListResourceSetRules;
-  export type Routes = purviewAccount.Routes;
-
-  export type AccountsGetAccountProperties200Response = responses.AccountsGetAccountProperties200Response;
-  export type AccountsGetAccountPropertiesdefaultResponse = responses.AccountsGetAccountPropertiesdefaultResponse;
-  export type AccountsUpdateAccountProperties200Response = responses.AccountsUpdateAccountProperties200Response;
-  export type AccountsUpdateAccountPropertiesdefaultResponse = responses.AccountsUpdateAccountPropertiesdefaultResponse;
-  export type AccountsGetAccessKeys200Response = responses.AccountsGetAccessKeys200Response;
-  export type AccountsGetAccessKeysdefaultResponse = responses.AccountsGetAccessKeysdefaultResponse;
-  export type AccountsRegenerateAccessKey200Response = responses.AccountsRegenerateAccessKey200Response;
-  export type AccountsRegenerateAccessKeydefaultResponse = responses.AccountsRegenerateAccessKeydefaultResponse;
-  export type CollectionsGetCollection200Response = responses.CollectionsGetCollection200Response;
-  export type CollectionsGetCollectiondefaultResponse = responses.CollectionsGetCollectiondefaultResponse;
-  export type CollectionsCreateOrUpdateCollection200Response = responses.CollectionsCreateOrUpdateCollection200Response;
-  export type CollectionsCreateOrUpdateCollectiondefaultResponse = responses.CollectionsCreateOrUpdateCollectiondefaultResponse;
-  export type CollectionsDeleteCollection204Response = responses.CollectionsDeleteCollection204Response;
-  export type CollectionsDeleteCollectiondefaultResponse = responses.CollectionsDeleteCollectiondefaultResponse;
-  export type CollectionsListCollections200Response = responses.CollectionsListCollections200Response;
-  export type CollectionsListCollectionsdefaultResponse = responses.CollectionsListCollectionsdefaultResponse;
-  export type CollectionsListChildCollectionNames200Response = responses.CollectionsListChildCollectionNames200Response;
-  export type CollectionsListChildCollectionNamesdefaultResponse = responses.CollectionsListChildCollectionNamesdefaultResponse;
-  export type CollectionsGetCollectionPath200Response = responses.CollectionsGetCollectionPath200Response;
-  export type CollectionsGetCollectionPathdefaultResponse = responses.CollectionsGetCollectionPathdefaultResponse;
-  export type ResourceSetRulesGetResourceSetRule200Response = responses.ResourceSetRulesGetResourceSetRule200Response;
-  export type ResourceSetRulesGetResourceSetRuledefaultResponse = responses.ResourceSetRulesGetResourceSetRuledefaultResponse;
-  export type ResourceSetRulesCreateOrUpdateResourceSetRule200Response = responses.ResourceSetRulesCreateOrUpdateResourceSetRule200Response;
-  export type ResourceSetRulesCreateOrUpdateResourceSetRuledefaultResponse = responses.ResourceSetRulesCreateOrUpdateResourceSetRuledefaultResponse;
-  export type ResourceSetRulesDeleteResourceSetRule200Response = responses.ResourceSetRulesDeleteResourceSetRule200Response;
-  export type ResourceSetRulesDeleteResourceSetRule204Response = responses.ResourceSetRulesDeleteResourceSetRule204Response;
-  export type ResourceSetRulesDeleteResourceSetRuledefaultResponse = responses.ResourceSetRulesDeleteResourceSetRuledefaultResponse;
-  export type ResourceSetRulesListResourceSetRules200Response = responses.ResourceSetRulesListResourceSetRules200Response;
-  export type ResourceSetRulesListResourceSetRulesdefaultResponse = responses.ResourceSetRulesListResourceSetRulesdefaultResponse;
-}
+import * as Models from "./models";
+import * as Pagination from "./paginateHelper";
+import * as Parameters from "./parameters";
+import * as Client from "./clientDefinitions";
+import * as Responses from "./responses";
+export { Models, Pagination, Parameters, Client, Responses };
+export { PurviewAccountClient } from "./purviewAccount";
