+++ sdk/purview/purview-administration-rest/src/account/parameters.ts
@@ -3,9 +3,9 @@
 
 import { RequestParameters } from "@azure-rest/core-client";
 import {
-  DataPlaneAccountUpdateParameters,
   AccessKeyOptions,
   Collection,
+  DataPlaneAccountUpdateParameters,
   ResourceSetRuleConfig,
 } from "./models";
