+++ sdk/purview/purview-administration-rest/src/index.ts
@@ -1,5 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from "./account";
-export * from "./metadataPolicies";
+import * as PurviewAccount from "./account";
+import * as PurviewMetadataPolicies from "./metadataPolicies";
+
+export { PurviewAccount };
+export { PurviewMetadataPolicies };
+export { PurviewAccountClient } from "./account/purviewAccount";
+export { PurviewMetadataPoliciesClient } from "./metadataPolicies/purviewMetadataPolicies";
